Comments and fixes, you may need to test on your PC to see if it works
I had to reinstall babel in order to use the import syntax you used. It may not work on your PC if you use linux.

I think we're probably going to demo on your computer because I dont have mongoDb installed locally, you could try to use the connection string to the online database instead but for demo probably on computer is better.